Go: Fix command groups and links (#2844)
* Fix command groups and links
